D1-Eshop-Database/requirements.txt
flask==3.0.0
pyodbc==5.0.1
python-dotenv==1.0.0
flask-cors==4.0.0